Add parameter to skip automatic Vault unseal (#11) (#12)
Add a `params.auxiliary_vault` parameter, which will cause the kit to bypass automatic vault unseal after a deploy. Previously, if an operator were to deploy a second Vault (not to be used for Genesis), an automatic unseal would fail because this kit was expecting to be deployed via a safe local --memory Vault backing it. Thus, the unseal logic would check for existing seal keys (which would be absent as the safe local --memory Vault does not contain any), and not attempt an automatic unseal. Worked great.

However, once that primary Vault is now deployed, an attempt to deploy a secondary Vault would fail during post-deploy execution as the kit would look for existing seal keys (which it would find because the primary Vault has seal keys) and go to unseal it with the seal keys of the primary Vault.

This functionality can now be skipped by setting `params.auxiliary_vault` to true (defaults to false). This is to be used with non-Genesis usage related Vaults. This fixes #11.